server/requirements.txt
django
djangorestframework
django-cors-headers 
python-dotenv
